完善一下结构，以及数据

+++ src/main/java/com/uifuture/springbootblockchain/pow/PowResult.java
@@ -6,24 +6,25 @@ package com.uifuture.springbootblockchain.pow;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * 工作量计算结果
  * @author chenhx
- * @version PowResult.java, v 0.1 2018-10-11 下午 9:18
+ * @version PowResult.java, v 0.1 2018-10-13 下午 4:50
  */
-
 @Data
 public class PowResult {
     /**
      * 计数器
      */
-    private long nonce;
+    private BigInteger nonce;
     /**
      * hash值
      */
     private String hash;
 
-    public PowResult(long nonce, String hash) {
+    public PowResult(BigInteger nonce, String hash) {
         this.nonce = nonce;
         this.hash = hash;
     }
